restructure mlir codegen folder
