+++ src/codegen/mlir/to_mlir/mod.rs
@@ -1,0 +1,4 @@
+pub mod types;
+
+// Re-export the trait and implementations
+pub use types::ToMlir;
